pyproject.toml
[project]
    dependencies    = ["asyncio >=3.4.3", "pillow >= 11.1.0", "websockets>=15.0", "aiohttp>=3.8.0"]
    description     = "Spectacles-2-Unitree Coordination Server"
    name            = "spectacles-2-unitree"
    readme          = "README.md"
    requires-python = ">=3.12"
    version         = "0.1.0"
